requirements.txt
Django==3.2.12
psycopg2-binary==2.9.3
djangorestframework==3.13.1
gunicorn==20.1.0
python-dotenv==0.19.2
django-filter==21.1